+++ src/lapidary/runtime/__init__.py
@@ -1,6 +1,7 @@
 __all__ = (
     'Body',
     'ClientBase',
+    'ClientError',
     'Cookie',
     'FormExplode',
     'HTTPError',
@@ -14,6 +15,7 @@ __all__ = (
     'Response',
     'Responses',
     'SecurityRequirements',
+    'ServerError',
     'Simple',
     'StatusCode',
     'UnexpectedResponseError',
@@ -30,6 +32,6 @@ from .annotations import Body, Cookie, Header, Metadata, Path, Query, Response, 
 from .client_base import ClientBase
 from .model import ModelBase
 from .model.encode_param import FormExplode, MatrixExplode, Simple
-from .model.error import HTTPError, UnexpectedResponseError
+from .model.error import ClientError, HTTPError, ServerError, UnexpectedResponseError
 from .operation import delete, get, head, patch, post, put, trace
 from .types_ import NamedAuth, SecurityRequirements